pyproject.toml
[project]
name = "OSUBeatmapRecommender"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "yaowan233", email = "572473053@qq.com"},
]
dependencies = [
    "sqlalchemy>=2.0.32",
    "sqlmodel>=0.0.21",
    "asyncpg>=0.29.0",
    "httpx>=0.27.0",
    "loguru>=0.7.2",
    "expiringdict>=1.2.2",
    "tqdm>=4.66.5",
    "rosu-pp-py>=1.1.0",
    "nonebot2>=2.3.3",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
